+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `module_build` attribute to RPM push items.
 - Added `rpm_filter_arch` parameter to errata source, to select a subset of RPMs
   by architecture.
+- Added `ModuleMdSourcePushItem` class. Source modulemd documents are now represented
+  by this class rather than `ModuleMdPushItem`.
 
 ## [2.7.0] - 2021-06-10